requirements.txt
Flask>=0.10,<1.0

Flask-SQLAlchemy>=2.0,<3.0

Flask-WTF>=0.11,<1.0

Flask-Login>=0.2.11,<1.0
Flask-Bcrypt>=0.6.2,<1.0
python-social-auth>=0.2,<1.0

pytest>=2.6.3
WebTest>=2.0,<3.0
factory_boy>=2.5,<3.0

Flask-Script
